Correcion de errores

+++ Promedio/src/Main.java
@@ -2,9 +2,9 @@ import javax.swing.*;
 
 public class Main {
     public static void main(String[] args) {
-        int numero1 =Integer.parseInt(JOptionPane.showInputDialog("Ingrese el primer digito: "));
-        int numero2 =Integer.parseInt(JOptionPane.showInputDialog("Ingrese el segundo digito: "));
-        int numero3 =Integer.parseInt(JOptionPane.showInputDialog("Ingrese el tercer digito: "));
+        float numero1 =Float.parseFloat(JOptionPane.showInputDialog("Ingrese el primer digito: "));
+        float numero2 =Float.parseFloat(JOptionPane.showInputDialog("Ingrese el segundo digito: "));
+        float numero3 =Float.parseFloat(JOptionPane.showInputDialog("Ingrese el tercer digito: "));
         float resultado = (numero1 + numero2 + numero3) / 3;
         JOptionPane.showMessageDialog(null, resultado);
